split User to another application
